% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_primer.R
\name{gen_ilp_single_primer}
\alias{gen_ilp_single_primer}
\title{Generate ILP for Single Primer Data}
\usage{
gen_ilp_single_primer(primer_data)
}
\arguments{
\item{primer_data}{List returned by \code{generate_single_primer_data}}
}
\value{
Result from \code{generate_multiplex_ilp_single_primers_direct}
}
\description{
Executes the ILP model generation for single primer data created by
\code{generate_single_primer_data}.
}
\examples{
\dontrun{
primer_data <- generate_single_primer_data(n_primers = 50, n_loci = 5)
ilp_result <- gen_ilp_single_primer(primer_data)
}

}
